+++ lib/loops/worker.rb
@@ -3,15 +3,14 @@ class Worker
   attr_reader :name
   attr_reader :pid
 
-  def initialize(name, logger, &blk)
+  def initialize(name, logger, engine, &blk)
     raise "Need a worker block!" unless block_given?
+    
     @name = name
     @logger = logger
-    
-    @pid = nil
-    @ppid = $$
-    
+    @engine = engine    
     @worker_block = blk
+    
     @shutdown = false
   end
   
@@ -21,10 +20,19 @@ class Worker
   
   def run
     return if shutdown?
-    @pid = Kernel.fork do
-      $0 = "loop worker: #{@name}\0"
-      @pid = Process.pid
-      @worker_block.call
+    if @engine == 'fork'
+      @pid = Kernel.fork do
+        $0 = "loop worker: #{@name}\0"
+        @pid = Process.pid
+        @worker_block.call
+        exit(0)
+      end
+    elsif @engine == 'thread'
+      @thread = Thread.start do
+        @worker_block.call
+      end
+    else
+      raise "Invalid engine name: #{@engine}"
     end
   rescue Exception => e
     logger.error("Exception from worker: #{e} at #{e.backtrace.first}")
@@ -32,22 +40,38 @@ class Worker
   
   def running?(verbose = false)
     return false if shutdown?
-    return false unless @pid
-    Process.waitpid(@pid, Process::WNOHANG)
-    res = Process.kill(0, @pid)
-    logger.debug("KILL(#{@pid}) = #{res}") if verbose
-    true
-  rescue Exception => e
-    logger.error("Exception from kill: #{e} at #{e.backtrace.first}") if verbose
-    false
+    if @engine == 'fork'
+      return false unless @pid
+      begin
+        Process.waitpid(@pid, Process::WNOHANG)
+        res = Process.kill(0, @pid)
+        logger.debug("KILL(#{@pid}) = #{res}") if verbose
+        return true
+      rescue Errno::ESRCH, Errno::ECHILD, Errno::EPERM => e
+        logger.error("Exception from kill: #{e} at #{e.backtrace.first}") if verbose
+        return false
+      end
+    elsif @engine == 'thread'
+      @thread && @thread.alive?
+    else
+      raise "Invalid engine name: #{@engine}"
+    end
   end
   
   def stop(force = false)
     @shutdown = true
-    sig = force ? 'SIGKILL' : 'SIGTERM'
-    logger.debug("Sending #{sig} to ##{@pid}")
-    Process.kill(sig, @pid)
-  rescue Exception => e
-    logger.error("Exception from kill: #{e} at #{e.backtrace.first}")
+    if @engine == 'fork'
+      begin
+        sig = force ? 'SIGKILL' : 'SIGTERM'
+        logger.debug("Sending #{sig} to ##{@pid}")
+        Process.kill(sig, @pid)
+      rescue Errno::ESRCH, Errno::ECHILD, Errno::EPERM=> e
+        logger.error("Exception from kill: #{e} at #{e.backtrace.first}")
+      end
+    elsif @engine == 'thread'
+      force && !defined?(::JRuby) ? @thread.kill! : @thread.kill
+    else
+      raise "Invalid engine name: #{@engine}"
+    end
   end
 end
